cdk/lambdas/requirements.txt
mysql-connector-python==8.3.0
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
urllib3==1.25.11

